fix ReSpec errors due to obsolete definitions
removal of obsolete definitions

ReSpec report:
Found definition for "3PL", but nothing links to it. This is usually a spec bug!
How to fix: You can do one of the following...

Add a class="lint-ignore" attribute the definition.
Either remove the definition or change <dfn> to another type of HTML element.
If you meant to export the definition, add class="export" to the definition.
To silence this warning entirely, set lint: { "no-unused-dfns": false } in your respecConfig.

Occurred 1 times at:

<dfn> element
(plugin: "core/linter-rules/no-unused-dfns")

Found definition for "Identity Proofing", but nothing links to it. This is usually a spec bug!
How to fix: You can do one of the following...

Add a class="lint-ignore" attribute the definition.
Either remove the definition or change <dfn> to another type of HTML element.
If you meant to export the definition, add class="export" to the definition.
To silence this warning entirely, set lint: { "no-unused-dfns": false } in your respecConfig.

Occurred 1 times at:

<dfn> element
(plugin: "core/linter-rules/no-unused-dfns")

Found definition for "NIST", but nothing links to it. This is usually a spec bug!
How to fix: You can do one of the following...

Add a class="lint-ignore" attribute the definition.
Either remove the definition or change <dfn> to another type of HTML element.
If you meant to export the definition, add class="export" to the definition.
To silence this warning entirely, set lint: { "no-unused-dfns": false } in your respecConfig.

Occurred 1 times at:

<dfn> element
(plugin: "core/linter-rules/no-unused-dfns")

Found definition for "TI", but nothing links to it. This is usually a spec bug!
How to fix: You can do one of the following...

Add a class="lint-ignore" attribute the definition.
Either remove the definition or change <dfn> to another type of HTML element.
If you meant to export the definition, add class="export" to the definition.
To silence this warning entirely, set lint: { "no-unused-dfns": false } in your respecConfig.

Occurred 1 times at:

<dfn> element