Update index.html
The current specification identifies exp in the VP as the expiration date to be checked. However, it was determined that the expirationDate in the VC contained within the VP is the canonical expiration date. A wallet provider could specify an incorrect exp that doesn't match the expirationDate. In the interest of performance and simplicity, propose that the expirationDate be checked directly, rather than adding a redundant check to confirm that the two values match.